Added ChatController
