+++ src/main/java/com/benjazzy/mixchat/controller/MixController.java
@@ -1,5 +1,6 @@
 package com.benjazzy.mixchat.controller;
 
+import com.benjazzy.mixchat.MixChat;
 import com.benjazzy.mixchat.MixUI;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -19,12 +20,6 @@ import java.util.concurrent.ExecutionException;
 public class MixController {
     private ConnectController connectController;
 
-    @FXML
-    private MenuItem connectMenu;           /** Menu item that contains the connect and disconect items. */
-    @FXML
-    private TextField message;              /** TextField containing the message the user wants to send. */
-    @FXML
-    private Button sendMessage;             /** Button to send message. */
     @FXML
     private Button connect;
     @FXML
@@ -69,6 +64,16 @@ public class MixController {
     }
 
     /**
+     * Disconnect from the chat.
+     *
+     * @param event
+     */
+    @FXML
+    private void Disconnect(ActionEvent event) {
+        MixUI.getInstance().getChat().disconnect();
+    }
+
+    /**
      * Connect to the chat in connectWindowField.
      *
      * @param channelName
@@ -88,55 +93,17 @@ public class MixController {
             loader.setLocation(getClass().getResource("/MixChatWindow.fxml"));
             try {
                 root = loader.load();
+
+                MixChat mixChat = new MixChat(root);
+
+                ChatController chatController = loader.getController();
+                chatController.setMixChat(mixChat);
+
+                chat.setContent(root);
+
+                mixChat.connect(channelName);
             } catch (IOException e) {
                 e.printStackTrace();
-            }
-
-            MixUI.getInstance().getChat().connect(channelName);
-        }
-    }
-
-    /**
-     * Disconnect from the chat.
-     *
-     * @param event
-     */
-    @FXML
-    private void Disconnect(ActionEvent event) {
-        MixUI.getInstance().getChat().disconnect();
-    }
-
-    /**
-     * Send the message in message to the Mixer chat.
-     *
-     * @param event
-     */
-    @FXML
-    private void SendMessage(ActionEvent event) {
-        String m = message.getText();
-        if (MixUI.getInstance().getChat().isConnected()) {
-
-            message.setText("");
-            MixUI.getInstance().getChat().sendMessage(m);
-        } else {
-            System.out.println("Error not connected");
-        }
-    }
-
-    /**
-     * Called when any key is pressed in message.  If the key is enter send message.
-     *
-     * @param event
-     */
-    @FXML
-    private void handleMessageEnter(KeyEvent event) {
-        if (event.getCode() == KeyCode.ENTER) {
-            String m = message.getText();
-            if (MixUI.getInstance().getChat().isConnected()) {
-                message.setText("");
-                MixUI.getInstance().getChat().sendMessage(m);
-            } else {
-                System.out.println("Error not connected");
             }
         }
     }
